add special handling for teachers
While we want to avoid sending all staff and only send staff with mapped roles,
we have to make an exception for staff who teach classes. This commit makes sure
any staff that teaches a section is included with their role set to teacher
(if they do not have a mapped role normally).